requirements.txt
discord.py
aiohttp
google-generativeai
PyNaCl


